Clean up redundant files and consolidate server
