+++ package.json
@@ -2,11 +2,10 @@
   "name": "geminispector",
   "version": "1.0.0",
   "description": "**多人格生成 + Gemini APIクロスチェックシステム**",
-  "main": "index.js",
+  "main": "server.js",
   "type": "commonjs",
   "scripts": {
-    "start": "node index.js",
-    "start:factcheck": "ts-node src/factcheck-system/server.ts",
+    "start": "node server.js",
     "build": "tsc",
     "test": "jest"
   },
