--- text-eval-api/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "text-eval-api",
-  "version": "1.0.0",
-  "description": "Simple API for text evaluation using Gemini",
-  "main": "index.js",
-  "scripts": {
-    "start": "node index.js"
-  },
-  "type": "commonjs",
-  "dependencies": {
-    "dotenv": "^17.0.0",
-    "express": "^5.1.0",
-    "node-fetch": "^2.7.0"
-  }
-}
